Ensure commiting Image Store Uploads honors the serverTimeout parameter. (#142)
I've noticed issues when uploading packages to a stand-alone cluster.
Occassionally, it will get a timeout error - even though the
serverTimeout value is set to a high enough value. I tracked it down to
the 'commit' stage. That code is not honoring the serverTimeout value.
Updating this locally resolved all our time-out issues when uploading
packages to our stand-alone SF Clusters.

This is my first time trying to push content on gitHub. I tried to
follow the instructions, but I apologize if I didn't do something
correct.

[https://github.com/microsoft/service-fabric/issues/1520](url)
#1520

Co-authored-by: Thomas Thatcher <Tom@enertia.biz>